new modify

+++ ViewModels/CategorySalesViewModel.cs
@@ -3,5 +3,5 @@
     public int CategoryId { get; set; }
     public string CategoryName { get; set; }
     public decimal TotalSales { get; set; }
-    public int ItemCount { get; set; }
+    public decimal ItemCount { get; set; }
 }
