+++ ViewModels/DailySalesViewModel.cs
@@ -4,8 +4,8 @@
     {
         public DateTime Date { get; set; }
         public int OrderCount { get; set; }
-        public decimal TotalSales { get; set; }
+        public double TotalSales { get; set; }
         public int CancelledCount { get; set; }
-        public decimal CancelledAmount { get; set; }
+        public double CancelledAmount { get; set; }
     }
 }
