+++ ViewModels/MonthlyReportViewModel.cs
@@ -5,7 +5,7 @@
         public int Month { get; set; }
         public int Year { get; set; }
         public List<DailySalesViewModel> DailySales { get; set; }
-        public decimal TotalSales { get; set; }
+        public double TotalSales { get; set; }
         public int TotalOrders { get; set; }
     }
 }
